commit-5
Gone through all four operations in the implementation. It seems all four execute well. However, there is still a bit of an issue with overtime and deductions in tax management, and taxAmount and taxPercentage in the financial record, which do not meet the expected output. Nonetheless, it should also be noted that it does not exceed any of the resultant logic. It remains stable.